// archive/sdk-pre-vnext-20260606/src/runtime/generated/runtime/v1/auth.ts
export enum AppMode {
    /**
     * @generated from protobuf enum value: APP_MODE_UNSPECIFIED = 0;
     */
    UNSPECIFIED = 0,
    /**
     * @generated from protobuf enum value: APP_MODE_LITE = 1;
     */
    LITE = 1,
    /**
     * @generated from protobuf enum value: APP_MODE_CORE_ONLY = 2;
     */
    CORE_ONLY = 2,
    /**
     * @generated from protobuf enum value: APP_MODE_FULL = 3;
     */
    FULL = 3
}
/**
 * @generated from protobuf enum nimi.runtime.v1.WorldRelation
 */
export enum WorldRelation {
    /**
     * @generated from protobuf enum value: WORLD_RELATION_UNSPECIFIED = 0;
     */
    UNSPECIFIED = 0,
    /**
     * @generated from protobuf enum value: WORLD_RELATION_NONE = 1;
     */
    NONE = 1,
    /**
     * @generated from protobuf enum value: WORLD_RELATION_RENDER = 2;
     */
    RENDER = 2,
    /**
     * @generated from protobuf enum value: WORLD_RELATION_EXTENSION = 3;
     */
    EXTENSION = 3
}