// archive/sdk-pre-vnext-20260606/src/runtime/generated/runtime/v1/common.ts
export interface ResourceSelectors {
    /**
     * @generated from protobuf field: repeated string conversation_ids = 1
     */
    conversationIds: string[];
    /**
     * @generated from protobuf field: repeated string message_ids = 2
     */
    messageIds: string[];
    /**
     * @generated from protobuf field: repeated string document_ids = 3
     */
    documentIds: string[];
    /**
     * @generated from protobuf field: map<string, string> labels = 10
     */
    labels: {
        [key: string]: string;
    };
}
// ...
export enum ExternalPrincipalType {
    /**
     * @generated from protobuf enum value: EXTERNAL_PRINCIPAL_TYPE_UNSPECIFIED = 0;
     */
    UNSPECIFIED = 0,
    /**
     * @generated from protobuf enum value: EXTERNAL_PRINCIPAL_TYPE_AGENT = 1;
     */
    AGENT = 1,
    /**
     * @generated from protobuf enum value: EXTERNAL_PRINCIPAL_TYPE_APP = 2;
     */
    APP = 2,
    /**
     * @generated from protobuf enum value: EXTERNAL_PRINCIPAL_TYPE_SERVICE = 3;
     */
    SERVICE = 3
}
// ...
class ResourceSelectors$Type extends MessageType<ResourceSelectors> {
    constructor() {
        super("nimi.runtime.v1.ResourceSelectors", [
            { no: 1, name: "conversation_ids", kind: "scalar", repeat: 2 /*RepeatType.UNPACKED*/, T: 9 /*ScalarType.STRING*/ },
            { no: 2, name: "message_ids", kind: "scalar", repeat: 2 /*RepeatType.UNPACKED*/, T: 9 /*ScalarType.STRING*/ },
            { no: 3, name: "document_ids", kind: "scalar", repeat: 2 /*RepeatType.UNPACKED*/, T: 9 /*ScalarType.STRING*/ },
            { no: 10, name: "labels", kind: "map", K: 9 /*ScalarType.STRING*/, V: { kind: "scalar", T: 9 /*ScalarType.STRING*/ } }
        ]);
    }
    create(value?: PartialMessage<ResourceSelectors>): ResourceSelectors {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.conversationIds = [];
        message.messageIds = [];
        message.documentIds = [];
        message.labels = {};
        if (value !== undefined)
            reflectionMergePartial<ResourceSelectors>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: ResourceSelectors): ResourceSelectors {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* repeated string conversation_ids */ 1:
                    message.conversationIds.push(reader.string());
                    break;
                case /* repeated string message_ids */ 2:
                    message.messageIds.push(reader.string());
                    break;
                case /* repeated string document_ids */ 3:
                    message.documentIds.push(reader.string());
                    break;
                case /* map<string, string> labels */ 10:
                    this.binaryReadMap10(message.labels, reader, options);
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    private binaryReadMap10(map: ResourceSelectors["labels"], reader: IBinaryReader, options: BinaryReadOptions): void {
        let len = reader.uint32(), end = reader.pos + len, key: keyof ResourceSelectors["labels"] | undefined, val: ResourceSelectors["labels"][any] | undefined;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case 1:
                    key = reader.string();
                    break;
                case 2:
                    val = reader.string();
                    break;
                default: throw new globalThis.Error("unknown map entry field for nimi.runtime.v1.ResourceSelectors.labels");
            }
        }
        map[key ?? ""] = val ?? "";
    }
    internalBinaryWrite(message: ResourceSelectors, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* repeated string conversation_ids = 1; */
        for (let i = 0; i < message.conversationIds.length; i++)
            writer.tag(1, WireType.LengthDelimited).string(message.conversationIds[i]);
        /* repeated string message_ids = 2; */
        for (let i = 0; i < message.messageIds.length; i++)
            writer.tag(2, WireType.LengthDelimited).string(message.messageIds[i]);
        /* repeated string document_ids = 3; */
        for (let i = 0; i < message.documentIds.length; i++)
            writer.tag(3, WireType.LengthDelimited).string(message.documentIds[i]);
        /* map<string, string> labels = 10; */
        for (let k of globalThis.Object.keys(message.labels))
            writer.tag(10, WireType.LengthDelimited).fork().tag(1, WireType.LengthDelimited).string(k).tag(2, WireType.LengthDelimited).string(message.labels[k]).join();
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message nimi.runtime.v1.ResourceSelectors
 */
export const ResourceSelectors = new ResourceSelectors$Type();