// archive/sdk-pre-vnext-20260606/src/runtime/generated/runtime/v1/grant.ts
export enum PolicyMode {
    /**
     * @generated from protobuf enum value: POLICY_MODE_UNSPECIFIED = 0;
     */
    UNSPECIFIED = 0,
    /**
     * @generated from protobuf enum value: POLICY_MODE_PRESET = 1;
     */
    PRESET = 1,
    /**
     * @generated from protobuf enum value: POLICY_MODE_CUSTOM = 2;
     */
    CUSTOM = 2
}
/**
 * @generated from protobuf enum nimi.runtime.v1.AuthorizationPreset
 */
export enum AuthorizationPreset {
    /**
     * @generated from protobuf enum value: AUTHORIZATION_PRESET_UNSPECIFIED = 0;
     */
    UNSPECIFIED = 0,
    /**
     * @generated from protobuf enum value: AUTHORIZATION_PRESET_READ_ONLY = 1;
     */
    READ_ONLY = 1,
    /**
     * @generated from protobuf enum value: AUTHORIZATION_PRESET_FULL = 2;
     */
    FULL = 2,
    /**
     * @generated from protobuf enum value: AUTHORIZATION_PRESET_DELEGATE = 3;
     */
    DELEGATE = 3
}